requirements.txt
Flask==2.3.0
Werkzeug==2.3.0
pymongo
dnspython
python-dotenv
nltk
requests
torch
transformers
certifi
beautifulsoup4
newspaper3k
trafilatura
lxml
itsdangerous
gunicorn
pandas
scikit-learn
matplotlib
seaborn
Flask-WTF
Flask-Limiter
Flask-Talisman
bleach
email-validator
cryptography
bcrypt
